Renamed SystemCapability App Extension filename and ExtractExtension function
