+++ src/components/application_manager/rpc_plugins/sdl_rpc_plugin/src/sdl_rpc_plugin.cc
@@ -79,7 +79,8 @@ void SDLRPCPlugin::OnApplicationEvent(
   }
 }
 
-void SDLRPCPlugin::DeleteSubscriptions(app_mngr::ApplicationSharedPtr app) {
+void SDLRPCPlugin::DeleteSubscriptions(
+    app_mngr::ApplicationSharedPtr app) {
   auto& ext = SystemCapabilityAppExtension::ExtractExtension(*app);
   auto subscriptions = ext.Subscriptions();
   for (auto& sub : subscriptions) {
